contract execution successfull

+++ src/utils/functions.ts
@@ -2,6 +2,7 @@ import * as vscode from "vscode";
 import * as fs from "fs";
 import path from "path";
 import { logger } from "../lib";
+import { ABIFragment } from "../types";
 
 export const createABIFile = (file: string) => {
   try {
@@ -14,14 +15,36 @@ export const createABIFile = (file: string) => {
     const fileData = fs.readFileSync(filePath, { encoding: "utf-8" });
     const fileName = file.substring(0, file.length - 5);
 
+    const abi: Array<ABIFragment> = JSON.parse(fileData).abi;
+
+    const abiFunctions = abi.filter((e) => e.type === "function");
+
+    const functionsValue = abiFunctions.map((func) => {
+      return {
+        ...func,
+        inputs: func.inputs.map((e) => {
+          return { ...e, value: "" };
+        }),
+      };
+    });
+
     if (!fs.existsSync(path.join(path_, "starkex", fileName))) {
       fs.mkdirSync(path.join(path_, "starkex", fileName));
     }
 
-    fs.writeFileSync(
-      path.join(path_, "starkex", fileName, `${fileName}_abi.json`),
-      JSON.stringify(JSON.parse(fileData).abi)
-    );
+    if (
+      !fs.existsSync(
+        path.join(path_, "starkex", fileName, `${fileName}_abi.json`)
+      )
+    ) {
+      fs.writeFileSync(
+        path.join(path_, "starkex", fileName, `${fileName}_abi.json`),
+        JSON.stringify(functionsValue)
+      );
+    } else {
+      logger.log(`${fileName}_abi.json already exist.`);
+    }
+
     logger.log("ABI file created successfully.");
   } catch (error) {
     logger.log(`Error while writing to file: ${error}`);
@@ -41,14 +64,23 @@ export const createAddressFile = (file: string) => {
       fs.mkdirSync(path.join(path_, "starkex", fileName));
     }
 
-    fs.writeFileSync(
-      path.join(path_, "starkex", fileName, `${fileName}_address.json`),
-      JSON.stringify({
-        name: fileName,
-        address: "",
-        classHash: "",
-      })
-    );
+    if (
+      !fs.existsSync(
+        path.join(path_, "starkex", fileName, `${fileName}_address.json`)
+      )
+    ) {
+      fs.writeFileSync(
+        path.join(path_, "starkex", fileName, `${fileName}_address.json`),
+        JSON.stringify({
+          name: fileName,
+          address: "",
+          classHash: "",
+        })
+      );
+    } else {
+      logger.log(`${fileName}_address.json already exist.`);
+    }
+
     logger.log("Address file created successfully.");
   } catch (error) {
     logger.log(`Error while writing to file: ${error}`);
